Adjust walltime and memory requests

+++ ush/narre/evs_get_prepbufr.sh
@@ -10,32 +10,42 @@ set -x
 
 modnam=$1
 
-
-
 if [ $modnam = prepbufr ] ; then
 
  mkdir -p $WORK/prepbufr.$VDATE
+ [[ ! -d $COMOUTsmall/prepbufr.${VDATE} ]] && mkdir -p $COMOUTsmall/prepbufr.${VDATE} 
  export output_base=${WORK}/pb2nc
 
  for vhr in 00  03  06  09  12  15  18   21  ; do
   if [ -s $COMINobsproc/rap.${VDATE}/rap.t${vhr}z.prepbufr.tm00 ] ; then 
-   split_by_subset $COMINobsproc/rap.${VDATE}/rap.t${vhr}z.prepbufr.tm00
-   for subset in ADPSFC SFCSHP MSONET; do
-    if [ -s ${WORK}/${subset} ]; then
-     cat ${WORK}/${subset} >> $WORK/prepbufr.$VDATE/rap.t${vhr}z.prepbufr.tm00
+    if [ ! -e $COMOUTsmall/prepbufr.${VDATE}/prepbufr.t${vhr}z.completed ] ; then 
+      split_by_subset $COMINobsproc/rap.${VDATE}/rap.t${vhr}z.prepbufr.tm00
+      for subset in ADPSFC SFCSHP MSONET; do
+        if [ -s ${WORK}/${subset} ]; then
+          cat ${WORK}/${subset} >> $WORK/prepbufr.$VDATE/rap.t${vhr}z.prepbufr.tm00
+        fi
+      done
+      export bufrpath=$WORK/prepbufr.$VDATE
+      for grid in G130 G242 ; do
+        export vbeg=${vhr}
+        export vend=${vhr}
+        export verif_grid=$grid
+        if [ -s $WORK/prepbufr.$VDATE/rap.t${vhr}z.prepbufr.tm00 ] ; then
+         ${METPLUS_PATH}/ush/run_metplus.py -c ${PARMevs}/metplus_config/machine.conf -c ${GRID2OBS_CONF}/Pb2nc_obsRAP_Prepbufr.conf
+         export err=$?; err_chk
+         cp ${WORK}/pb2nc/prepbufr_nc/*.nc $WORK/prepbufr.${VDATE}
+       fi
+     done
+       if [ -s ${WORK}/pb2nc/prepbufr_nc/prepbufr.t${vhr}z.G*.nc ] ; then
+         cp ${WORK}/pb2nc/prepbufr_nc/prepbufr.t${vhr}z.G*.nc $WORK/prepbufr.${VDATE}
+	 cp ${WORK}/pb2nc/prepbufr_nc/prepbufr.t${vhr}z.G*.nc $COMOUTsmall/prepbufr.${VDATE}
+	 >$COMOUTsmall/prepbufr.${VDATE}/prepbufr.t${vhr}z.completed  
+       fi
+    else 
+     #Get file from restart: 
+     cp $COMOUTsmall/prepbufr.${VDATE}/prepbufr.t${vhr}z.G*.nc $WORK/prepbufr.${VDATE}
     fi
-   done
-   export bufrpath=$WORK/prepbufr.$VDATE
-   for grid in G130 G242 ; do
-     export vbeg=${vhr}
-     export vend=${vhr}
-     export verif_grid=$grid
-     if [ -s $WORK/prepbufr.$VDATE/rap.t${vhr}z.prepbufr.tm00 ] ; then
-       ${METPLUS_PATH}/ush/run_metplus.py -c ${PARMevs}/metplus_config/machine.conf -c ${GRID2OBS_CONF}/Pb2nc_obsRAP_Prepbufr.conf
-       export err=$?; err_chk
-       cp ${WORK}/pb2nc/prepbufr_nc/*.nc $WORK/prepbufr.${VDATE}
-     fi
-   done
+
   else 
     echo "WARNING:  No Prepbufr data available $COMINobsproc/rap.${VDATE}/rap.t${vhr}z.prepbufr.tm00 for ${VDATE}"
     if [ $SENDMAIL = YES ] ; then
@@ -48,11 +58,6 @@ if [ $modnam = prepbufr ] ; then
     fi
   fi
  done
- #Save for restart
-  if [ -d $WORK/prepbufr.${VDATE} ] ; then
-      cp -r $WORK/prepbufr.${VDATE}  $COMOUTsmall
-  fi
- 
 
 fi
